Removes unused redisson CommandExecutor import
Hi Vladimir,

There is an unused import on line 32 (org.redisson.command.CommandExecutor). 

I suspect this to be a typo becasue:
1) the import org.redisson.command.CommandExecutor is not referenced anywhere inside the class definition
2) the class definition (SpringDataRedisBasedProxyManager) is a spring-data implementation of the ProxyManager. I would not suspect to see references to vendor-specific implementations here because spring data redis already provides an abstraction.

My builds still run but my IDE is complaining because I don't have redisson on my classpath (this is surely the case for others as well, spring-boot-data-starter-redis only supplies jedis and lettuce).